Move all of this logic out of the POCO and into a separate class.
