+++ src/SparkPost/SparkPost.csproj
@@ -95,6 +95,7 @@
     <Compile Include="UpdateRecipientListResponse.cs" />
     <Compile Include="UpdateSendingDomainResponse.cs" />
     <Compile Include="Utilities\Jsonification.cs" />
+    <Compile Include="Utilities\MailMessageMapping.cs" />
     <Compile Include="Utilities\SnakeCase.cs" />
     <Compile Include="ValueMappers\AnonymousValueMapper.cs" />
     <Compile Include="ValueMappers\BooleanValueMapper.cs" />
